fix: gate Bugsnag debug page to non-production + untrack supabase/.temp
- Bugsnag test page returns 404 in production to prevent quota abuse
- Added supabase/.temp/ to .gitignore and removed tracked CLI state files
