--- supabase/.temp/linked-project.json
@@ -1,1 +1,0 @@
-{"ref":"ijsdpozjfjjufjsoexod","name":"Banco-de-Dados - MassuerMacht","organization_id":"ybcvqggiorccvprvbjwo","organization_slug":"ybcvqggiorccvprvbjwo"}
