--- supabase/.temp/pooler-url
@@ -1,1 +1,0 @@
-postgresql://postgres.ijsdpozjfjjufjsoexod@aws-1-us-east-1.pooler.supabase.com:5432/postgres
