--- supabase/.temp/postgres-version
@@ -1,1 +1,0 @@
-17.6.1.032
