--- supabase/.temp/rest-version
@@ -1,1 +1,0 @@
-v13.0.5
